requirements.txt
streamlit
pytube
python-dotenv
validators
moviepy
ffmpeg-python